+++ packages/health_connector_hk_ios/ios/health_connector_hk_ios/Sources/health_connector_hk_ios/mappers/health_record_mappers/ExerciseTypeMapper.swift
@@ -233,7 +233,7 @@ extension HKWorkoutActivityType {
             .running
         case .walking:
             .walking
-        case .cycling:
+        case .cycling, .cyclingStationary:
             .cycling
         case .hiking:
             .hiking